"Adding auth0 controller file "
