+++ api/index.js
@@ -9,7 +9,7 @@ const { Categories, Products, Categories_Products } = conn.models;
 
 const port= process.env.PORT || 3001;
 // Syncing all the models at once.
-conn.sync({ force: false }).then(() => {
+conn.sync({ force: true }).then(() => {
   server.listen(port, async () => {
     const products = await Products.findAll()
     if(!products.length){
