+++ api/src/asset/AdminEmail.JSON
@@ -1,0 +1,36 @@
+[
+    { 
+        "email": "yones.rpy@gmail.com"
+        
+    },
+    
+    {
+        "email": "cesarv793@gmail.com"
+    },
+    
+    {
+         
+        "email": "gimemezgaston@gmail.com"
+
+    },
+    {
+         
+        "email": "martinignacioojeda@gmail.com"
+
+    },
+    {
+         
+        "email": "banmugo2910@gmail.com"
+
+    },
+    {
+         
+        "email": "mastropierro.francisco@gmail.com"
+
+    },
+    {
+         
+        "email": "vlasovacristina@gmail.com"
+
+    }
+]
